requirements.txt
sentence-transformers
psycopg2-binary
langgraph
langchain
langchain_community
regex
numpy
scikit-learn
google-generativeai
